Move DesignerSerializerAttribute to S.CM.Primitives and add to StringDictionary
