+++ src/libraries/System.Collections.Specialized/src/System.Collections.Specialized.csproj
@@ -18,6 +18,7 @@
   </ItemGroup>
   <ItemGroup>
     <Reference Include="System.Collections.NonGeneric" />
+    <Reference Include="System.ComponentModel.Primitives" />
     <Reference Include="System.Runtime" />
     <Reference Include="System.Runtime.Extensions" />
     <Reference Include="System.Threading" />
