+++ src/libraries/System.Collections.Specialized/src/System/Collections/Specialized/StringDictionary.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 
+using System.ComponentModel.Design.Serialization;
+
 namespace System.Collections.Specialized
 {
     /// <devdoc>
@@ -12,6 +14,8 @@ namespace System.Collections.Specialized
     /// </devdoc>
     [Serializable]
     [System.Runtime.CompilerServices.TypeForwardedFrom("System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089")]
+    [DesignerSerializer("System.Diagnostics.Design.StringDictionaryCodeDomSerializer, System.Design, Version=4.0.0.0, PublicKeyToken=b03f5f7f11d50a3a",
+                        "System.ComponentModel.Design.Serialization.CodeDomSerializer, System.Design, Version=4.0.0.0, PublicKeyToken=b03f5f7f11d50a3a")]
     public class StringDictionary : IEnumerable
     {
         // For compatibility, we want the Keys property to return values in lower-case.
