+++ src/libraries/System.ComponentModel.Primitives/src/System.ComponentModel.Primitives.csproj
@@ -15,6 +15,7 @@
     <Compile Include="System\ComponentModel\DesignerSerializationVisibility.cs" />
     <Compile Include="System\ComponentModel\DesignerSerializationVisibilityAttribute.cs" />
     <Compile Include="System\ComponentModel\DesignOnlyAttribute.cs" />
+    <Compile Include="System\ComponentModel\Design\Serialization\DesignerSerializerAttribute.cs" />
     <Compile Include="System\ComponentModel\DisplayNameAttribute.cs" />
     <Compile Include="System\ComponentModel\EventHandlerList.cs" />
     <Compile Include="System\ComponentModel\IComponent.cs" />
