+++ src/libraries/System.ComponentModel.TypeConverter/src/System.ComponentModel.TypeConverter.csproj
@@ -220,7 +220,6 @@
     <Compile Include="System\ComponentModel\Design\Serialization\ContextStack.cs" />
     <Compile Include="System\ComponentModel\Design\Serialization\DefaultSerializationProviderAttribute.cs" />
     <Compile Include="System\ComponentModel\Design\Serialization\DesignerLoader.cs" />
-    <Compile Include="System\ComponentModel\Design\Serialization\DesignerSerializerAttribute.cs" />
     <Compile Include="System\ComponentModel\Design\Serialization\SerializationStore.cs" />
     <Compile Include="System\ComponentModel\Design\Serialization\IDesignerLoaderHost.cs" />
     <Compile Include="System\ComponentModel\Design\Serialization\IDesignerLoaderService.cs" />
